@ v29: 킬스위치/강제청산 버그·수익방어·레짐집행·진입개선·전략랩 스플릿·파킹정리 (2026-07-08)
크리티컬 버그:
- 킬스위치 오발동(잔고델타→미실현손실율) + checkDrawdownShield 강제청산(해외전용 판정)
- 국내 스마트집행 더블매수(smartFill 스킵)
- SQL 타입버그 3곳(RECONCILER $4::numeric, MOMENTUM_HOLD $2::int/$3::float)
- 인버스 3중 이중매수(BUY dedup) + 파킹 개별주 강제매수 제거

수익방어/진입:
- 반복손절 하드차단 + 스마트재진입 삭제(hard-gates + smart-reentry.ts 삭제 + 단위테스트)
- Connors RSI2 근거 진입가드(MA20아래 보류, MA60위+과매도 예외)
- 손절 하드플로어 -6/-7→-5/-6, 레짐 적응형 집행, 장초반 갭SL 스위프

운영:
- 라이브 인버스 허용+HOLD가드, 뉴스 테마 폴백+Groq TPD 서킷브레이커
- SEC 버튼 캐시우선+DB영구화(마이그124), 전략랩 클린스플릿(해외 strategy_bucket)
- DB-KIS 프로액티브 정합(3분+야간15분)

@

+++ src/ai/track-b/momentum-hold.ts
@@ -193,7 +193,7 @@ export async function setMomentumHoldMeta(chainId: string, count: number, peak: 
     await safeQuery(
       `UPDATE transaction_chains
        SET metadata = COALESCE(metadata, '{}'::jsonb)
-         || jsonb_build_object('momentum_hold_count', $2, 'momentum_hold_last_peak', $3)
+         || jsonb_build_object('momentum_hold_count', $2::int, 'momentum_hold_last_peak', $3::float)
        WHERE id = $1`,
       [chainId, count, peak],
     );
